For MP4, use AAC decoder config buffer directly.

+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/natives/aac/AacDecoder.java
@@ -60,7 +60,7 @@ public class AacDecoder extends NativeResourceHolder {
 
     long buffer = 0;
     for (int i = 0; i < config.length; i++) {
-      buffer |= config[i] << (i << 3);
+      buffer |= ((long) config[i]) << (i << 3);
     }
 
     configureRaw(buffer);
